trees_and_graphs/test_avl.py
import pytest

from avl import Node, BST
# Tests of methods in avl.py. Supporting both the Node and BST classes.

# Tests for Node Class
def test_node_construction():
    n = Node('key', 'value')
    assert n.key == 'key'
    assert n.value == 'value'
    assert n.left_child == None 
    assert n.right_child == None
    assert n.parent == None
    #assert n.height == 0




# Tests for BST Class
@pytest.fixture(scope='class')
def empty_bst():
    return BST()

@pytest.fixture(scope='class')
def rooted_bst():
    bst = BST()
    bst.put(6, 'A')
    return bst
    
def test_bst_construction():
    bst = BST()
    assert bst.root == None
    assert bst.size == 0

class TestPut:

    @pytest.mark.parametrize('new_node, parent, root_bf', 
                                [
                                    (Node(3, 'B'), 'A', 1),
                                    (Node(9, 'C'), 'A', 1),
                                    (Node(2, 'D'), 'B', 2),
                                    (Node(4, 'E'), 'B', 2),
                                    (Node(10, 'F'), 'C', 2)
                                ])
    def test__put(self, rooted_bst, new_node, parent, root_bf):
        bst = rooted_bst
        bst._put(new_node, bst.root)
        assert new_node.parent.value == parent
        assert rooted_bst.root.balance_factor == root_bf

    def test_put_root(self, empty_bst):
        assert empty_bst.root == None
        assert empty_bst.size == 0
        empty_bst.put(6, 'val')
        assert empty_bst.root != None
        assert empty_bst.size == 1

    @pytest.mark.parametrize('key, value, size', 
                                [
                                    (3, 'B', 2),
                                    (9, 'C', 3),
                                    (2, 'D', 4),
                                    (4, 'E', 5)
                                ])
    def test_put(self, empty_bst, key, value, size):
        empty_bst.put(key, value)
        assert empty_bst.size == size

class TestBalance:

    right_heavy_right_child_right_heavy_bst = [
                (64, 'A'), (32, 'B'), (96, 'C'), (16, 'D'), (48, 'E'),
                (80, 'F'), (112, 'G'), (88, 'H'), (104, 'I'), (120, 'J'),
                (116, 'K')
            ]
    
    right_heavy_right_child_left_heavy_bst = [
                (64, 'A'), (32, 'B'), (96, 'C'), (16, 'D'), (48, 'E'),
                (80, 'F'), (112, 'G'), (72, 'H'), (88, 'I'), (120, 'J'),
                (84, 'K')
            ]
  


    left_heavy_left_child_left_heavy_bst = [
                (64, 'A'), (32, 'B'), (96, 'C'), (16, 'D'), (48, 'E'),
                (80, 'F'), (112, 'G'), (8, 'H'), (24, 'I'), (56, 'J'),
                (20, 'K')
            ]


    left_heavy_left_child_right_heavy_bst = [
                (64, 'A'), (32, 'B'), (96, 'C'), (16, 'D'), (48, 'E'),
                (80, 'F'), (112, 'G'), (8, 'H'), (40, 'I'), (56, 'J'),
                (52, 'K')
            ]

    def build_unbalanced_tree(self, node_list):
        bst = BST()
        for node in node_list:
            new_node = Node(node[0], node[1])
            if bst.root == None:
                bst.root = new_node
            else:
                current_node = bst.root
                while current_node:
                    if new_node.key <= current_node.key:
                        if current_node.left_child == None:
                            new_node.parent = current_node
                            current_node.left_child = new_node
                            break
                        else:
                            current_node = current_node.left_child
                    else:
                        if current_node.right_child == None:
                            new_node.parent = current_node
                            current_node.right_child = new_node
                            break
                        else:
                            current_node = current_node.right_child
        return bst

    

    @pytest.mark.parametrize('parent_node, parent_bf, left_node, left_node_bf, right_node, right_node_bf',
                        [
                            (Node(50, 'A'), 0, Node(25, 'B'), 5, Node(75, 'C'), 5),
                            (Node(50, 'A'), 2, Node(25, 'B'), 3, Node(75, 'C'), 5),
                            (Node(50, 'A'), 2, Node(25, 'B'), 5, Node(75, 'C'), 3),
                            (Node(50, 'A'), 6, None, None, Node(75, 'C'), 5),
                            (Node(50, 'A'), 6, Node(25, 'B'), 5, None, None)
                        ])
    def test_update_balance_factor(self, parent_node, parent_bf, left_node, left_node_bf, right_node, right_node_bf):
        bst = BST()
        bst.root = parent_node
        parent_node.left_child = left_node
        parent_node.right_child = right_node

        if left_node:
            left_node.balance_factor = left_node_bf
        if right_node:
            right_node.balance_factor = right_node_bf

        assert bst.update_balance_factor(bst.root) == parent_bf

        
    def populate_bst(self, bst, node_list):
        for node in node_list:
            #print(f'adding: {node[1]}')
            bst.put(node[0], node[1])
    
    @pytest.mark.parametrize('node_list',
                        [
                            (right_heavy_right_child_right_heavy_bst),
                            (right_heavy_right_child_left_heavy_bst),
                            (left_heavy_left_child_left_heavy_bst),
                            (left_heavy_left_child_right_heavy_bst)
                        ])
    def test_rebalance(self, node_list):
        # All test trees are unbalacned about root.
        bst = BST()
        self.populate_bst(bst, node_list)
        print(bst.size)
        #bst.rebalance(bst.root)
        #assert abs(bst.root.left_child.height - bst.root.right_child.height) <= 1
        assert 1 == 1





